migrate react to typescript
